Scilab Task Part B

+++ test/trans.c
@@ -1,10 +1,10 @@
 #include<stdlib.h>
 
-int **trans(int **mat, int m, int n) {
+double **trans(double **mat, int m, int n) {
     int i, j;
-    int **x = malloc(n*sizeof(int*));
+    double **x = malloc(n*sizeof(double*));
     for(i = 0; i < n; i++) {
-        x[i] = malloc(m*sizeof(int));
+        x[i] = malloc(m*sizeof(double));
     }
     for(i = 0; i < m; i++) {
         for(j = 0; j < n; j++) {
